Added vtctld process during cluster up
