+++ dev.env
@@ -73,6 +73,17 @@ PATH=$(prepend_path "$PATH" "$VTROOT/dist/chromedriver")
 PATH=$(prepend_path "$PATH" "$VTROOT/dist/node/bin")
 export PATH
 
+# Etcd path.
+case $(uname) in
+  Linux)  etcd_platform=linux;;
+  Darwin) etcd_platform=darwin;;
+esac
+
+ETCD_VERSION=$(cat "${VTROOT}/dist/etcd/.installed_version")
+ETCD_BINDIR="${VTROOT}/dist/etcd/etcd-${ETCD_VERSION}-${etcd_platform}-amd64/"
+PATH=$(prepend_path "$PATH" "$ETCD_BINDIR")
+export PATH
+
 # GOROOT sanity
 go_bin=$(which go)
 go_env=$(go env | grep GOROOT | cut -f 2 -d\")
