+++ go/vt/vttest/clustertest/etcd_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package clustertest
 
 import (
-	"net/http"
 	"testing"
 )
 
@@ -25,17 +24,4 @@ func TestEtcdServer(t *testing.T) {
 	testURL(t, "http://localhost:2379/v2/keys", "generic etcd url")
 	testURL(t, "http://localhost:2379/v2/keys/vitess/global", "vitess global key")
 	testURL(t, "http://localhost:2379/v2/keys/vitess/zone1", "vitess zone1 key")
-}
-
-func testURL(t *testing.T, url string, testCaseName string) {
-	statusCode := getStatusForURL(url)
-	if got, want := statusCode, 200; got != want {
-		t.Errorf("select:\n%v want\n%v for %s", got, want, testCaseName)
-	}
-}
-
-// getStatusForUrl returns the status code for the URL
-func getStatusForURL(url string) int {
-	resp, _ := http.Get(url)
-	return resp.StatusCode
 }
